pyproject.toml
[tool.poetry]
name = "demo-python-finance"
version = "0.1.0"
description = ""
authors = ["natthasath <natthasath.sak@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "demo_python_finance"}]

[tool.poetry.dependencies]
python = "^3.10"
yfinance = "^0.2.31"
requests = "^2.31.0"
lxml = "^4.9.3"
matplotlib = "^3.8.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
